pull utils into y-axis split
